+++ lib/exporto_api.rb
@@ -15,11 +15,15 @@ module ExportoAPI
   autoload :RateLimitError, "exporto_api/errors"
   autoload :ServerError, "exporto_api/errors"
   autoload :AuthResource, "exporto_api/resources/auth_resource"
+  autoload :LabelResource, "exporto_api/resources/label_resource"
   autoload :LabelMethodResource, "exporto_api/resources/label_method_resource"
+  autoload :ReturnShipmentResource, "exporto_api/resources/return_shipment_resource"
   autoload :ShipmentResource, "exporto_api/resources/shipment_resource"
 
   module Objects
+    autoload :LabelResponse, "exporto_api/objects/label_response"
     autoload :LabelMethodResponse, "exporto_api/objects/label_method_response"
+    autoload :ReturnShipmentResponse, "exporto_api/objects/return_shipment_response"
     autoload :ShipmentResponse, "exporto_api/objects/shipment_response"
     autoload :TokenResponse, "exporto_api/objects/token_response"
   end