+++ lib/exporto_api/objects/return_shipment_response.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ExportoAPI
+  module Objects
+    class ReturnShipmentResponse < Base; end
+  end
+end